Jupyter Notebook DWWP Research
This is the original research notebook, composed in Google Colab. As this project was for my final Capstone with Thinkful, there was a time limit on its completion. It is not "done". There will be additional statistical tests that were not taught in the coursework added. I am exploring them and learning about them this week, and will determine which ones best suit this data, which has no normal distribution. The pearson r correlation tests are complete.